Continued working on the carrer screen
