+++ components/BlogHerosection.jsx
@@ -6,7 +6,7 @@ export default function BlogHerosection() {
     <div>
       <div className="bg-[#033A08]">
         <div className="blog-bg_image">
-          <div className="xl:px-[100px] lg:px-[75px] md:px-[50px] sm:px-[35px] px-4 py-[21px] bg-[#171717B2]   xl:py-[120px] sm:py-[40px] md:py-[75px] lg:py-[100px]">
+          <div className="xl:px-[100px] lg:px-[75px] md:px-[50px] sm:px-[35px] px-4 py-[21px] bg-[#171717B2] xl:py-[120px] sm:py-[40px] md:py-[75px] lg:py-[100px]">
             <h1 className="font-semibold xl:text-[48px] lg:text-[44px]  text-2xl sm:text-4xl md:text-6xl lg:mb-4 sm:mb-2 mb-1 text-[#FEFCFC] font-rope">
               <span className="text-warning600">Payyme</span> in the news
             </h1>
