--- components/Layout.jsx
@@ -1,13 +1,0 @@
-import Image from "next/image";
-import Navbar from "./Navbar";
-import Footer from "./Footer";
-
-export default function Layout({ children }) {
-  return (
-    <>
-      <Navbar />
-      <main>{children}</main>
-      <Footer />
-    </>
-  );
-}
