+++ components/career/CareerMission.jsx
@@ -8,7 +8,7 @@ export default function CareerMission() {
     },
     {
       id: 2,
-      imgSrc: "/asset/team-1_1.png",
+      imgSrc: "/asset/team-1_2.png",
     },
     {
       id: 3,
@@ -21,10 +21,10 @@ export default function CareerMission() {
   ];
 
   return (
-    <div className="px-[100px] py-[45px] bg-[#FFFFFF]">
-      <div className="flex gap-[95px] items-center">
-        <div className="w-full justify-center">
-          <div className="mb-[156px]">
+    <div className="xl:px-[100px] lg:px-[65px] md:px-[48px] sm:px-8 px-5 py-[45px] bg-[#FFFFFF]">
+      <div className="lg:flex xl:gap-[95px] gap-[48px]">
+        <div className="w-full justify-center flex flex-col">
+          <div className="sm:mb-auto mb-8">
             <h1 className="font-medium xl:text-5xl sm:text-xl md:text-2xl lg:text-3xl text-base md:mb-4 mb-2 text-[#043623] text-center">
               <span style={{ width: "29px", borderTop: "3.5px solid #0DA168" }}>
                 Our
@@ -58,14 +58,17 @@ export default function CareerMission() {
             </h6>
           </div>
         </div>
-        <div className="w-full grid grid-cols-2 gap-6">
+        <div className="w-full grid grid-cols-2 md:gap-6 sm:gap-4 gap-2 mt-[44px] lg:mt-0">
           {TeamList.map((team) => (
             <Image
               src={team.imgSrc}
               height={292}
               width={292}
               alt="Team member icon"
-              className="object-cover"
+              className="object-cover w-full"
+              style={{
+                borderRadius: "0px 57.02px 0px 57.02px",
+              }}
             />
           ))}
         </div>
