+++ components/career/Role.jsx
@@ -3,22 +3,26 @@ import Image from "next/image";
 export default function CareerRole() {
   return (
     <>
-      <div className="rounded-[40px] bg-[#032015] mx-10 text-[#FEFCFC] font-medium">
+      <div className="rounded-[40px] bg-[#032015] lg:mx-10 mx-5 md:mx-7 text-[#FEFCFC] font-medium mb-[60px]">
         <h3
-          className="pb-4 pt-[59px] ps-[64px] text-5xl"
+          className="pb-4 pt-[59px] l:px-[64px] lg:px-[48px] md:px-[36px] sm:px-[24px] px-4 md:text-5xl text-2xl"
           style={{ borderBottom: "1px solid #A9A7A7" }}
         >
           Find your best role
         </h3>
         <div>
           <div
-            className="px-[64px] pt-6 pb-[40px] flex"
+            className="xl:px-[64px] lg:px-[48px] md:px-[36px] sm:px-[24px] px-4 pt-6 pb-[40px] md:flex"
             style={{ borderBottom: "1px solid #A9A7A7" }}
           >
-            <h5 className="text-2xl">Business & Customer Operation</h5>
-            <div className="p-4 flex ms-auto w-[713px] justify-between bg-[#ECF3F54D] rounded-[10px]">
+            <h5 className="md:text-2xl text-xl w-full mb-6 md:mb-0">
+              Business & Customer Operation
+            </h5>
+            <div className="p-4 flex ms-auto justify-between bg-[#ECF3F54D] rounded-[10px] w-full">
               <div>
-                <h5 className="text-2xl mb-1">Customer Success Manager</h5>
+                <h5 className="md:text-2xl text-xl mb-1">
+                  Customer Success Manager
+                </h5>
                 <p className="font-normal text-sm text-[#D3D0D0]">
                   Full-Time Role
                 </p>
@@ -34,14 +38,18 @@ export default function CareerRole() {
             </div>
           </div>
           <div
-            className="px-[64px] pt-6 pb-[40px] flex"
+            className="xl:px-[64px] lg:px-[48px] md:px-[36px] sm:px-[24px] px-4 pt-6 pb-[40px] md:flex"
             style={{ borderBottom: "1px solid #A9A7A7" }}
           >
-            <h5 className="text-2xl">Engineering</h5>
-            <div className="ms-auto">
-              <div className="p-4 flex ms-auto w-[713px] justify-between mb-6 bg-[#ECF3F54D] rounded-[10px]">
+            <h5 className="md:text-2xl text-xl w-full mb-6 md:mb-0">
+              Engineering
+            </h5>
+            <div className="ms-auto w-full">
+              <div className="p-4 flex ms-auto w-full justify-between mb-6 bg-[#ECF3F54D] rounded-[10px]">
                 <div>
-                  <h5 className="text-2xl mb-1">Head of Engineering</h5>
+                  <h5 className="md:text-2xl text-xl mb-1">
+                    Head of Engineering
+                  </h5>
                   <p className="font-normal text-sm text-[#D3D0D0]">
                     Full-Time Role
                   </p>
@@ -57,9 +65,11 @@ export default function CareerRole() {
                   </div>
                 </div>
               </div>
-              <div className="p-4 flex ms-auto w-[713px] justify-between bg-[#ECF3F54D] rounded-[10px]">
+              <div className="p-4 flex ms-auto w-full justify-between bg-[#ECF3F54D] rounded-[10px]">
                 <div>
-                  <h5 className="text-2xl mb-1">Senior Software Engineering</h5>
+                  <h5 className="md:text-2xl text-xl mb-1">
+                    Senior Software Engineering
+                  </h5>
                   <p className="font-normal text-sm text-[#D3D0D0]">
                     Full-Time Role
                   </p>
@@ -78,13 +88,15 @@ export default function CareerRole() {
             </div>
           </div>
           <div
-            className="px-[64px] pt-6 pb-[40px] flex"
+            className="xl:px-[64px] lg:px-[48px] md:px-[36px] sm:px-[24px] px-4 pt-6 pb-[40px] md:flex"
             style={{ borderBottom: "1px solid #A9A7A7" }}
           >
-            <h5 className="text-2xl">Finance</h5>
-            <div className="p-4 flex ms-auto w-[713px] justify-between bg-[#ECF3F54D] rounded-[10px]">
+            <h5 className="md:text-2xl text-xl w-full mb-6 md:mb-0">Finance</h5>
+            <div className="p-4 flex ms-auto w-full justify-between bg-[#ECF3F54D] rounded-[10px]">
               <div>
-                <h5 className="text-2xl mb-1">Senior Software Engineering</h5>
+                <h5 className="md:text-2xl text-xl mb-1">
+                  Senior Software Engineering
+                </h5>
                 <p className="font-normal text-sm text-[#D3D0D0]">
                   Full-Time Role
                 </p>
