+++ pages/faq/index.jsx
@@ -1,6 +1,6 @@
 import Faq from "@/components/faq/Faq";
 import FaqOptions from "@/components/faq/FaqOptions";
-import Layout from "@/components/Layout";
+import Layout from "@/layout";
 
 export default function Blog() {
   return (
