+++ pages/how/index.jsx
@@ -1,0 +1,19 @@
+import HowSection from "@/components/HowSection";
+import HowSub from "@/components/HowSubSection";
+
+import VideoComponent from "@/components/PayymeVideo";
+import Layout from "@/layout";
+
+export default function How() {
+  return (
+    <Layout>
+      <div className="bg-[#FAF8FE] navbar-margin">
+        <div className="mb-[80px]">
+          <HowSection />
+          <HowSub />
+          <VideoComponent />
+        </div>
+      </div>
+    </Layout>
+  );
+}
